apply this fix:
From: Andrew Aye <AndrewA@terminalreality.com>
To: ode@q12.org
Date: Wed, 25 Feb 2004 10:14:09 -0600
Subject: [ODE] ODE Change

Fixes some issues that arrise when disabled/enabled pairs interact.  Bug was
probably introduced when the variable inits were commented out ( which is
the other possible fix - just make sure the matrices are init to 0 - this
other way keeps with the intent to write to the variable as few times as
possible ).

 - Andrew


git-svn-id: https://opende.svn.sourceforge.net/svnroot/opende/trunk/ode@604 685f7672-210f-0410-90b4-fe3ad19314fe